Add Material Design Animations and concepts
The idea is to migrate the hole app to Material Design guidelines in order to create a better user experience. New animations were created for Fragment-Fragment transitions, RecyclerView bind views enter and leaving the Add Repos Fragment with a explode animation. The search view was added to the toolbar and recyclerview swipe to refresh as well. An Image Dialog was added in order to show the image bigger in front on the screen. Several XMLs with new animations were also added. All material design colors are now usable on the app with their correct values.